[FIX] l10n_cr_edi_sale_coupon: Avoid error calculate percentage again
Before this commit, the discount percentage was being calculate always
using the value of the fixed discount, even if the fixed discount was
actually calculated from the percentage in the first place. This made
decimal loss because the rounding to two decimals.
If the order does not have coupons or promotions it does not need to
recalculate the percentage discount, so now it doesn't.
Issue#17486